Time: 1 ms (93.27%), Space: 45.1 MB (22.88%) - LeetHub

+++ 0152-maximum-product-subarray/0152-maximum-product-subarray.java
@@ -1,15 +1,17 @@
 class Solution {
     public int maxProduct(int[] nums) {
         int n = nums.length;
-        double pre = 1, suff = 1, ans = Integer.MIN_VALUE;
+        int pre = 1, suff = 1, ans = Integer.MIN_VALUE;
+
         for(int i=0; i<n; i++){
             pre = pre * nums[i];
             suff = suff * nums[n-i-1];
             ans = Math.max(ans, Math.max(pre, suff));
-            
+
             if(pre == 0) pre = 1;
             if(suff == 0) suff = 1;
         }
-        return (int)ans;
+        
+        return ans;
     }
 }
